Refactor some code to put legend in a different file
