+++ src/styles/modal.scss
@@ -14,7 +14,7 @@
 
 .modal-content {
   background-color: #FCFAF6;
-  padding: 20px 35px;
+  padding: 20px 50px;
   border-radius: 14px;
   width: 30%;
   text-align: left;
@@ -94,7 +94,7 @@
 
 
 #info-content {
-  padding: 20px;
+  padding: 20px 50px;
 
     ul li {
       padding: 10px 0;
